% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBrainW1.R
\docType{data}
\name{data_brain_w1}
\alias{data_brain_w1}
\title{The human brain exon array dataset, time window 1}
\format{A list with multiple elements}
\description{
A subset of 1,000 genes are included for demonstration purpose.
}
\details{
\itemize{
  \item X the N by p data matrix, N is the number of samples and p is the number of genes.
  \item Y the N by q confounder matrix (See implementation details in the user's guide).
  \item Yid labels for the individuals, left and right hemispheres from the same donors are treated as different individuals (See implementation details in the user's guide).
  \item regions labels for the brain regions
  \item hemispheres '1' represents left hemisphere, '3' represents right hemisphere
  \item donor_labs labels for the donors
}
}
\examples{
load_all()
data(data_brain_w1)
}
\keyword{datasets}

