% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim5.R
\docType{data}
\name{data_example5}
\alias{data_example5}
\title{Simulated data example 5}
\format{A list with multiple elements}
\description{
Example 5 is similar to example 2, except that the true loadings are sparse. The variables with non-zero loadings are entries 301-400.
}
\details{
\itemize{
  \item X the N by p data matrix, number of samples=30, number of variables=400.
  \item Y the N by q confounder matrix, q=3, corresponding to three biological groups.
}
}
\examples{
load_all()
data(data_example5)
}
\keyword{datasets}

